refactor(paths): reorganize project structure and update requires toreflect new core/fact/parser strategies directories
