+++ lib/string_calculator/core/calculator.rb
