+++ lib/string_calculator/factories/component_factory.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-require_relative 'parser/delimiter_parser'
-require_relative 'parser/input_cleaner'
-require_relative 'parser/string_parser'
-require_relative 'calculator'
+require_relative '../parser/delimiter_parser'
+require_relative '../parser/input_cleaner'
+require_relative '../parser/string_parser'
+require_relative '../core/calculator'
 
 module StringCalculator
   class ComponentFactory
