+++ lib/string_calculator/parser/strategies/cleaning_strategy.rb
