+++ lib/string_calculator/parser/strategies/custom_delimiter_cleaning_strategy.rb
