+++ lib/string_calculator/parser/strategies/custom_delimiter_strategy.rb
