+++ lib/string_calculator/parser/strategies/default_cleaning_strategy.rb
