+++ lib/string_calculator/parser/strategies/default_delimiter_strategy.rb
