+++ lib/string_calculator/parser/strategies/delimiter_strategy.rb
